Structural Changes
Made module structural changes, fixed some common conventions issues, and renamed some files.
